backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Project Equilibrium — FastAPI backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "celery>=5.6.3",
    "cryptography>=46.0.7",
    "email-validator>=2.3.0",
    "fastapi>=0.135.3",
    "google-generativeai>=0.8.6",
    "groq>=1.1.2",
    "httpx>=0.28.1",
    "passlib[bcrypt]>=1.7.4",
    "pgvector>=0.4.2",
    "psycopg[binary,pool]>=3.3.3",
    "pydantic-settings>=2.13.1",
    "pydantic[email]>=2.13.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.26",
    "redis>=7.4.0",
    "slowapi>=0.1.9",
    "spacy>=3.8.14",
    "sqlalchemy>=2.0.49",
    "uvicorn[standard]>=0.44.0",
]
